refactor modular test suite file layout
